FlowClock initial commit
A smoothly animated clock for horizontal viewing on a docked iPhone.

This version is a proof of concept, feeling out the design. The intent
is to avoid pulsing lights or jumping images, while allowing the time
to be easily read.

Ultimately, I'd like this to be my default view in Standby mode, with
complications for date, day of week, first event of the new day, night
mode, etc.