Clean up
Stopped repeating myself so much and cleaned things up with a little
more decomposition. Now we can easily declare a new "Dial()" and have
it construct all its parts. Nicely, we can apply a color to the view,
so as long as we don't want e.g. some tick marks to be a different
color than others, we don't have to pass colors as parameters.

I also figured out why I was hitting the dynamic island; the offset
that I'm using to overlap the dials is also pushing the seconds dial
outside the safe area to the right. A simple fix is adding padding
to the leading edge that's the same size as the offset.