Add Dark Mode colors
Uses Colorsets to switch to a red scheme for night adapted eyes in Dark Mode.

Now, when Dark Mode is enabled, the dials will be dark red, with bright red
indicators for hours and minutes, and black indicators for seconds.

The idea here is that the seconds dial will continue to be an accent color
that deemphasizes it over the more significant values.

I would prefer to have a red display when Nightshift takes effect, too, but
for now, Dark Mode provides an easy way to test the color scheme.