Add 'protect sleep' alarm control
This introduces a three-state control for modifying your sleep schedule;
unfortunately, it doesn't do anything yet; I'm not sure it's possible to
access the sleep goal and alarm settings.

What I'd like this to do is toggle between your default alarm time (based
on your sleep goal and default bedtime as set via the Clock app), a custom
alarm time that is automatically set from your sleep goal and the current
time (when the button is pushed), and no alarm at all.

Additionally, the wake time should be shown, maybe as another dial that
shows sleep time remaining in hours.

I'm not fully confident of this UI, because it seems to me that you should
be able to set the wake time based on the current time, but then avoid
altering it if you bump the button accidentally. Maybe it needs a flyout
to confirm resetting the alarm.